structure tweaks
